fix : 별점 계산 및 타입 수정

+++ server/src/main/java/gohome/dailydaily/domain/product/entity/Product.java
@@ -37,7 +37,7 @@ public class Product extends BaseTime {
     private Integer price;
 
     @Column(nullable = false)
-    private Float score;
+    private Integer score;
 
     @Column(nullable = false)
     private Integer sale;
@@ -71,7 +71,7 @@ public class Product extends BaseTime {
         this.orderProducts.addAll(List.of(orderProducts));
     }
 
-    public void setScore(Float score) {
+    public void setScore(Integer score) {
         this.score = score;
     }
